+++ CHANGELOG.md
@@ -6,12 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.1.1] / 2023-10-03
 ### Features
+- `TestDiscoverer` enable log feature.
+### Updated
 - Update `ricaun.NUnit` to `1.3.1`
 ### Command
 - Update TargetFramework added `net5`
 - Update `JsonExtension` with `System.Text.Json`
 ### Console
 - Update `ricaun.Revit.Installation` to `1.1.0` (Fix: #30)
+### TestAdapter
+- Update `TestDiscoverer` with `Log` feature
+- Update `RevitTestConsole` with `RunTestReadWithLog` feature
 ### Fixed
 - Fix Folder not exist (Fix: #30)